refactor: rename billing entry view to index.vue
